Add mockups for login/logout/signup.
Using bootstrap to style new pages, unfortunately this screws things
up a little bit in some of the other components on the page. Will have
to cannibalize the parts of bootstrap that I need so the things that I
don't need will not contaminate the rest of the page.

Some other minor edits made to reflect nsides dev database and the new
mongo server IP address.