Fix api.
The app now renders the plot like expected. One of the biggest issues
was that I didn't have authentication information for the Amazon RDS
MySQL database; once I reset that and configured it properly the API
then functioned correctly.

The second issue was that by default the app retrieves waaaaay too
many results if many results are available for a drug. This is
probably desirable, but it's a major computational bottleneck and we
will have to figure out how to appropriately optimize. I changed it to
retrieve the top 10 outcomes instead.

The final issue may be a direct result of the previous issue---for
Fingolimod (my test drug, in the dev database), the graph renders as
blank for many outcomes... Perhaps we need to retrieve all results :/