Integrate agave platform/api.
Much of the code modifications in this commit/branch are adapted from
Choonhan Youn's example code in the job_api/ directory. Basically,
login/logout routes are replaced with the appropriate callbacks to the
agave authentication pages. This essentially invalidates all
login/logout code that has been included until this point, but that's
ok. We have also added a new template for submitting new jobs, but you
need to authenticate with Agave before you are able to access this
page.

Still to do:
- Hook up main app to login/submit routes
- Wire up the job submission page
- Make a custom loader for the flask cli utility
- Migrate everything to production (e.g., get production tokens and
hook in the server certificates)
- Switch away from nsides_dev database
- Lose dependency on mysql